refactor: move user to v2
