+++ src/app/favorite/service.rs
@@ -2,8 +2,8 @@ use crate::app::article::model::{Article, FetchBySlugAndAuthorId};
 use crate::app::article::service::{fetch_article, FetchArticle};
 use crate::app::favorite::model::{CreateFavorite, DeleteFavorite, Favorite, FavoriteInfo};
 use crate::app::tag::model::Tag;
-use crate::app::user::model::User;
 use crate::appv2::features::profile::entities::Profile;
+use crate::appv2::features::user::entities::User;
 use crate::error::AppError;
 use diesel::pg::PgConnection;
