+++ src/app/mod.rs
@@ -4,4 +4,3 @@ pub mod favorite;
 pub mod follow;
 pub mod healthcheck;
 pub mod tag;
-pub mod user;
