--- src/app/user/mod.rs
@@ -1,4 +1,0 @@
-pub mod api;
-pub mod model;
-pub mod request;
-pub mod response;
