+++ src/appv2/features/user/controllers.rs
@@ -1,18 +1,18 @@
-use super::model::{UpdateUser, User};
-use super::{request, response::UserResponse};
+use super::entities::{UpdateUser, User};
+use super::{presenters::UserResponse, requests};
 use crate::appv2::drivers::middlewares::auth;
 use crate::appv2::drivers::middlewares::state::AppState;
 use crate::utils::api::ApiResponse;
 use actix_web::{web, HttpRequest, HttpResponse};
 
-pub async fn signin(state: web::Data<AppState>, form: web::Json<request::Signin>) -> ApiResponse {
+pub async fn signin(state: web::Data<AppState>, form: web::Json<requests::Signin>) -> ApiResponse {
     let conn = &mut state.get_conn()?;
     let (user, token) = User::signin(conn, &form.user.email, &form.user.password)?;
     let res = UserResponse::from((user, token));
     Ok(HttpResponse::Ok().json(res))
 }
 
-pub async fn signup(state: web::Data<AppState>, form: web::Json<request::Signup>) -> ApiResponse {
+pub async fn signup(state: web::Data<AppState>, form: web::Json<requests::Signup>) -> ApiResponse {
     let conn = &mut state.get_conn()?;
     let (user, token) = User::signup(
         conn,
@@ -34,7 +34,7 @@ pub async fn me(req: HttpRequest) -> ApiResponse {
 pub async fn update(
     state: web::Data<AppState>,
     req: HttpRequest,
-    form: web::Json<request::Update>,
+    form: web::Json<requests::Update>,
 ) -> ApiResponse {
     let conn = &mut state.get_conn()?;
     let current_user = auth::get_current_user(&req)?;
