+++ src/appv2/features/user/mod.rs
@@ -1,1 +1,5 @@
+pub mod controllers;
+pub mod entities;
+pub mod presenters;
 pub mod repositories;
+pub mod requests;
