+++ src/appv2/features/user/repositories.rs
@@ -1,6 +1,6 @@
 use crate::app::follow::model::{CreateFollow, DeleteFollow, Follow};
-use crate::app::user::model::User;
 use crate::appv2::features::profile::entities::Profile;
+use crate::appv2::features::user::entities::User;
 use crate::error::AppError;
 use crate::utils::db::DbPool;
